requirements.txt
openai>=1.0.0
streamlit
python-dotenv
pydantic
agents @ git+https://github.com/openagents/openagents.git
